fix: use u64 transfer expiry

+++ crates/clipplus-transport/src/file_transfer.rs
@@ -10,12 +10,12 @@ pub enum TransferState {
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct FileTransferRequest {
     pub transfer_id: String,
-    pub expires_after_minutes: u32,
+    pub expires_after_minutes: u64,
     pub state: TransferState,
 }
 
 impl FileTransferRequest {
-    pub fn new_for_test(transfer_id: impl Into<String>, expires_after_minutes: u32) -> Self {
+    pub fn new_for_test(transfer_id: impl Into<String>, expires_after_minutes: u64) -> Self {
         Self {
             transfer_id: transfer_id.into(),
             expires_after_minutes,
@@ -23,7 +23,7 @@ impl FileTransferRequest {
         }
     }
 
-    pub fn is_expired_at_minute(&self, elapsed_minutes: u32) -> bool {
+    pub fn is_expired_at_minute(&self, elapsed_minutes: u64) -> bool {
         elapsed_minutes > self.expires_after_minutes
     }
 }
